+++ client/public/css/styles.css
@@ -526,4 +526,22 @@ nav .icon-burger .line {
   }
 }
 
-/* **********LANDING ********
+/* **********LANDING ******** */
+.travel-banner{
+  position: absolute;
+  right: 0;
+  width: 100%;
+  min-height: 100vh;
+  padding: 100px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  background: #111;
+  transition: 0.5s;
+}
+
+.property-container{
+  margin-top : 15vh;
+  margin-left: 10vh;
+  margin-right : 10vh;
+}
