+++ client/src/components/Landing.js
@@ -1,22 +1,16 @@
 import React from "react";
-import travel from "../images/travel.jpg";
-import Listings from "./Listings";
-import LoginButton from "./Authentication/LoginButton";
-import LogoutButton from "./Authentication/LogoutButton";
-import Properties from "./Properties";
+import PropertyTypeList from "./PropertyTypeList";
+import travel from "../images/travel.jpg"
 
 const Landing = () => {
+
   return (
     <div>
-      <div>
-        <section className="showcase">
-          <img src={travel} alt="suitcases" />
-        </section>
-        <Properties />
-      </div>
 
+      <PropertyTypeList />
     </div>
-  );
-};
+  )
+
+}
 
 export default Landing;
